refactor: Modalを構成要素ごとに切り分ける
