+++ frontend/src/shared/components/Modal/index.ts
@@ -1,0 +1,4 @@
+export * from "./Modal";
+export * from "./ModalHeader";
+export * from "./ModalBody";
+export * from "./ModalFooter";
